+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-- **Add items here** for the next release (Replace this comment)
+- **Service Worker Reload Loop Fix**: Fixed again
 
 # v0.1.4a